Path Stroker: Path/Generator Adapter is Now Workable
- old vcgen_adaptor and conv_stroke removed;
- path_generator_adapter is now able to process multi-segmented paths, but some additional coding is required to implement rewind and extended path commands.

+++ misc/libraries/aggx/aggx_shorten_path.h
@@ -55,7 +55,7 @@ namespace aggx
 			}
 			else
 			{
-				n = vs.size() - 1;
+				n = static_cast<int>(vs.size() - 1);
 				vertex_type& prev = vs[n-1];
 				vertex_type& last = vs[n];
 				d = (prev.dist - s) / prev.dist;
